requirements.txt
fastapi
python-dotenv
azure-ai-inference
pytube
uvicorn
youtube-transcript-api
pydantic
python-multipart
faster-whisper